Move everything from __init__.py to functions.py
